Add support for new `chrome` SVG argument in PTG
There is a bit of a hack included in this update: it seems that
MkDocs will insert a `<p></p>` tag before any non-code content in
tabbed layout. Because of this, when exporting without chrome we have
to manually set `margin-top: -1em` to align it to the top.

Not sure if this can or even needs to be worked around, however.